refactor extract stitching into it's own module
